Support regexp in "on" property for choregraphies

+++ test/cookbooks/test/recipes/default.rb
@@ -7,6 +7,7 @@
 # Need to clean up in case of multiple convergence tests
 execute 'Clean up machine' do
   command "rm /tmp/not_converging.tmp"
+  only_if "test -f /tmp/not_converging.tmp"
 end
 
 # Using chef provided resource
@@ -32,12 +33,25 @@ test_simple_resource 'not_converging' do
   only_if { false }
 end
 
+log "a_simple_log"
+log "another_log"
+
+custom_resource 'my converging custom resource'
+
+custom_resource 'my useless custom resource' do
+  only_if { false }
+end
 
 choregraphie 'execute' do
   on 'execute[converging]'
   on 'execute[not_converging]'
   on 'test_simple_resource[converging]'
   on 'test_simple_resource[not_converging]'
+  on 'custom_resource[my converging custom resource]'
+  on 'custom_resource[my useless custom resource]'
+
+  on /^log\[/
+
   before do |resource|
     Chef::Log.warn("I am called before! for resource " + resource.to_s)
     filename = resource.to_s.gsub(/\W+/, '_').gsub(/_$/,'')
@@ -48,3 +62,5 @@ choregraphie 'execute' do
     Chef::Log.warn('I am called at the end')
   end
 end
+
+log "a_log_defined_after_choregraphie"
